Book cards, filter, new areas

+++ Kutuphane/Controllers/BooksController.cs
@@ -48,17 +48,10 @@ namespace Kutuphane.Controllers
         // GET: Books/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
-            var book = await _context.Books
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (book == null)
-            {
-                return NotFound();
-            }
+            var book = await _context.Books.FirstOrDefaultAsync(m => m.Id == id);
+            if (book == null) return NotFound();
 
             return View(book);
         }
@@ -70,11 +63,9 @@ namespace Kutuphane.Controllers
         }
 
         // POST: Books/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Title,Author,Genre,Status,Notes")] Book book)
+        public async Task<IActionResult> Create([Bind("Id,Title,Author,Genre,Status,Notes,CoverImageUrl,Rating,PageCount,IsFavorite,Progress")] Book book)
         {
             if (ModelState.IsValid)
             {
@@ -88,30 +79,20 @@ namespace Kutuphane.Controllers
         // GET: Books/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var book = await _context.Books.FindAsync(id);
-            if (book == null)
-            {
-                return NotFound();
-            }
+            if (book == null) return NotFound();
+
             return View(book);
         }
 
         // POST: Books/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Title,Author,Genre,Status,Notes")] Book book)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Title,Author,Genre,Status,Notes,CoverImageUrl,Rating,PageCount,IsFavorite,Progress")] Book book)
         {
-            if (id != book.Id)
-            {
-                return NotFound();
-            }
+            if (id != book.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -122,14 +103,8 @@ namespace Kutuphane.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!BookExists(book.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!BookExists(book.Id)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -139,17 +114,10 @@ namespace Kutuphane.Controllers
         // GET: Books/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
-            var book = await _context.Books
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (book == null)
-            {
-                return NotFound();
-            }
+            var book = await _context.Books.FirstOrDefaultAsync(m => m.Id == id);
+            if (book == null) return NotFound();
 
             return View(book);
         }
@@ -160,10 +128,7 @@ namespace Kutuphane.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var book = await _context.Books.FindAsync(id);
-            if (book != null)
-            {
-                _context.Books.Remove(book);
-            }
+            if (book != null) _context.Books.Remove(book);
 
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
